+++ atcoder-problems-backend/src/lib.rs
@@ -1,0 +1,4 @@
+#[macro_use]
+extern crate diesel;
+
+pub(crate) mod sql;
