+++ atcoder-problems-backend/src/sql/schema.rs
